requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
rembg==2.0.67
onnxruntime==1.18.0
pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-image==0.22.0
scipy==1.12.0
PyMatting==1.1.8
filelock==3.13.1
typing_extensions==4.12.2
requests==2.32.3
